Refactor code to create separate component for overview
